@@ -1,42 +1,39 @@
 {
-  "name": "astro-syntax-docs",
+  "name": "docs",
   "type": "module",
   "version": "0.0.1",
   "scripts": {
     "dev": "astro dev",
     "start": "astro dev",
-    "build": "astro build",
+    "build": "astro check && astro build",
     "preview": "astro preview",
     "prettify": "prettier --write src",
     "astro": "astro"
   },
   "dependencies": {
-    "@astrojs/prefetch": "^0.2.1",
-    "@headlessui/react": "^1.7.14",
-    "clsx": "^1.2.1",
-    "preact": "^10.13.2",
+    "@astrojs/check": "^0.3.3",
+    "@astrojs/mdx": "^2.0.1",
+    "@astrojs/preact": "^3.0.1",
+    "@astrojs/react": "^3.0.7",
+    "@astrojs/tailwind": "^5.0.3",
+    "@headlessui/react": "^1.7.17",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
+    "astro": "^4.0.6",
+    "preact": "^10.19.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "tailwindcss": "^3.3.2",
-    "zustand": "^4.3.8"
+    "tailwindcss": "^3.4.0",
+    "typescript": "^5.3.3",
+    "zustand": "^4.4.7"
   },
   "devDependencies": {
-    "@astrojs/mdx": "^0.19.1",
-    "@astrojs/preact": "^2.1.0",
-    "@astrojs/react": "^2.1.3",
-    "@astrojs/tailwind": "^3.1.2",
-    "@tailwindcss/typography": "^0.5.9",
-    "@types/react": "^18.2.6",
-    "@types/react-dom": "^18.2.4",
-    "astro": "^2.4.5",
-    "astro-auto-import": "^0.3.0",
-    "astro-mdx-code-blocks": "^0.0.3",
-    "prettier": "^2.8.8",
-    "prettier-plugin-astro": "^0.8.1",
-    "prettier-plugin-organize-imports": "^3.2.2",
-    "prettier-plugin-tailwindcss": "^0.2.8",
-    "rehype-external-links": "^2.1.0",
-    "shiki": "0.14.2",
-    "typescript": "^5.0.4"
+    "@tailwindcss/typography": "^0.5.10",
+    "astro-mdx-code-blocks": "^0.0.6",
+    "clsx": "^2.0.0",
+    "prettier-plugin-astro": "^0.12.2",
+    "prettier-plugin-organize-imports": "^3.2.4",
+    "prettier-plugin-tailwindcss": "^0.5.9",
+    "rehype-external-links": "^3.0.0"
   }
 }
