@@ -25,7 +25,7 @@ export const Article: FunctionComponent<PropsWithChildren<ArticleProps>> = ({
         // pre
         "prose-pre:rounded-xl prose-pre:bg-slate-900 prose-pre:shadow-lg dark:prose-pre:bg-slate-800/60 dark:prose-pre:shadow-none dark:prose-pre:ring-1 dark:prose-pre:ring-slate-300/10",
         // hr
-        "dark:prose-hr:border-slate-800"
+        "dark:prose-hr:border-slate-800",
       )}
       {...otherProps}
     >
