@@ -18,7 +18,7 @@ export const Navigation: FunctionComponent<NavigationProps> = ({ className, curr
             <h2 className="font-display font-medium text-slate-900 dark:text-white">{ucfirst(section)}</h2>
             <ul
               role="list"
-              className="mt-2 space-y-2 border-l-2 border-slate-100 dark:border-slate-800 lg:mt-4 lg:space-y-4 lg:border-slate-200"
+              className="mt-2 space-y-2 border-l-2 border-slate-100 lg:mt-4 lg:space-y-4 lg:border-slate-200 dark:border-slate-800"
             >
               {links.map((link) => (
                 <li key={link.href} className="relative">
@@ -29,7 +29,7 @@ export const Navigation: FunctionComponent<NavigationProps> = ({ className, curr
                       "block w-full pl-3.5 before:pointer-events-none before:absolute before:-left-1 before:top-1/2 before:h-1.5 before:w-1.5 before:-translate-y-1/2 before:rounded-full",
                       link.href === currentPage
                         ? "font-semibold text-sky-500 before:bg-sky-500"
-                        : "text-slate-500 before:hidden before:bg-slate-300 hover:text-slate-600 hover:before:block dark:text-slate-400 dark:before:bg-secondary-400 dark:hover:text-secondary-200"
+                        : "text-slate-500 before:hidden before:bg-slate-300 hover:text-slate-600 hover:before:block dark:text-slate-400 dark:before:bg-secondary-400 dark:hover:text-secondary-200",
                     )}
                   >
                     {link.title}
